+++ src/main/java/com/charlie/morerecipes/repositories/RecipeRepository.java
@@ -3,5 +3,5 @@ package com.charlie.morerecipes.repositories;
 import com.charlie.morerecipes.domain.Recipe;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RecipeRepository extends CrudRepository<Recipe,Integer> {
+public interface RecipeRepository extends CrudRepository<Recipe,Long> {
 }
