+++ src/test/java/com/charlie/morerecipes/converters/CategoryCommandToCategoryTest.java
@@ -10,7 +10,7 @@ import static org.junit.Assert.*;
 public class CategoryCommandToCategoryTest {
 
     public static final String DESCRIPTION = "description";
-    public static final Integer INTEGER_VALUE = 1;
+    public static final Long INTEGER_VALUE = 1L;
 
     CategoryCommandToCategory converter;
