+++ src/test/java/com/charlie/morerecipes/converters/IngredientToIngredientCommandTest.java
@@ -12,7 +12,7 @@ import static org.junit.Assert.*;
 
 public class IngredientToIngredientCommandTest {
 
-    public static final Integer INTEGER_VALUE = 1;
+    public static final Long INTEGER_VALUE = 1L;
     public static final BigDecimal BIG_DECIMAL_VALUE = BigDecimal.valueOf(30l);
     public static final String DESCRIPTION = "description";
